pyproject.toml
[project]
name = "my-audio-summaries"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "gradio>=5.5.0",
    "ipykernel>=6.29.5",
    "loguru>=0.7.2",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.5",
    "promptic>=0.7.7",
    "pydantic>=2.9.2",
    "pypdf>=5.0.1",
    "pytesseract>=0.3.13",
    "pytest>=8.3.4",
    "sentry-sdk>=2.16.0",
    "tenacity>=9.0.0",
    "uvicorn>=0.32.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
